Backend/requirements.txt
asgiref==3.7.2
blinker==1.6.2
click==8.1.6
dj-database-url==2.1.0
Django==4.2.5
django-cors-headers==4.3.0
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
packaging==23.2
protobuf==4.21.12
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
PyQt3D==5.15.5
PyQt5==5.15.7
PyQt5-sip==12.11.0
PyQtChart==5.15.6
PyQtDataVisualization==5.15.5
PyQtNetworkAuth==5.15.5
PyQtPurchasing==5.15.5
PyQtWebEngine==5.15.6
pytz==2023.3.post1
SQLAlchemy==2.0.20
sqlparse==0.4.4
typing_extensions==4.7.1
Werkzeug==2.3.7
whitenoise==6.6.0